chore: rename module to include it vanity list
